# expr: Get rid of clap for performance reasons