As a side effect, we also now have the right behavior when adding arguments after `--help` or `--version`. Before this change, clap would have printed the help, but it is expected that flags are ignored if argc > 2 (more than 1 arg is given) 

Benchmarks:

```
$ export N=2500

# Before
$ hyperfine -m 1000 -- "./expr_clap $(yes 1 | head -n $N | paste -s - 2>&1 | sed 's/\t/ + /g')"
  Time (mean ± σ):      47.4 ms ±   3.3 ms    [User: 34.4 ms, System: 12.8 ms]
  Range (min … max):    37.3 ms …  74.4 ms    1000 runs

#After
$ hyperfine -m 1000 -- "./expr_no_clap $(yes 1 | head -n $N | paste -s - 2>&1 | sed 's/\t/ + /g')"
  Time (mean ± σ):      24.7 ms ±   2.2 ms    [User: 13.9 ms, System: 10.7 ms]
  Range (min … max):    18.8 ms …  34.7 ms    1000 runs
```