@RenjiSann 2025-02-24
by running the GNU tests manually, I realized that I forgot to handle `--` which can appear in the command. I made the necessary changes ([diff](https://github.com/uutils/coreutils/compare/d70956e5c6e6377bc150c54f82501eb6d83f9a76..b0d7ea414fd9e61516648ff8c3fd427dc222d59b))